UI_BE_Tech_Pro/requirement.txt
asgiref==3.2.10
Django==3.0
djangorestframework==3.12.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
